+++ moondream-ggml/core/CMakeLists.txt
@@ -18,7 +18,15 @@ set(BUILD_SHARED_LIBS OFF CACHE BOOL "Build ggml as a static library" FORCE)
 add_subdirectory(./dependencies/ggml)
 
 
-set(SOURCES src/moondream.cpp src/unicode_data.cpp src/unicode.cpp)
+set(SOURCES 
+    src/moondream.cpp
+    src/lm.cpp
+    src/mmproj.cpp
+    src/helpers.cpp
+    src/stb_image.cpp
+    src/unicode_data.cpp 
+    src/unicode.cpp
+)
 
 # Library target.
 add_library(moondream_ggml STATIC ${SOURCES})
@@ -29,9 +37,7 @@ target_include_directories(moondream_ggml PRIVATE ./dependencies/single_header/i
 target_compile_definitions(moondream_ggml PRIVATE MOONDREAM_LIBRARY_BUILD)
 
 # Executable target.
-# We only want to use stb_image for testing the executable, 
-# it should be omitted from the library target.
-add_executable(moondream_ggml_exe ${SOURCES} ./src/stb_image.cpp)
+add_executable(moondream_ggml_exe ${SOURCES})
 target_link_libraries(moondream_ggml_exe PRIVATE ggml)
 target_include_directories(moondream_ggml_exe PRIVATE ./dependencies/ggml/include)
 target_include_directories(moondream_ggml_exe PRIVATE ./dependencies/single_header/include)
